requirements.txt
asgiref==3.8.1
boto3==1.38.19
botocore==1.38.19
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.2.1
django-admin-honeypot @ git+https://github.com/dmpayton/django-admin-honeypot@develop
django-allauth==65.3.1
django-appconf==1.1.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-phonenumber-field==8.0.0
django-select2==8.3.0
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
phonenumbers==8.13.54
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
pytz==2025.1
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.3.0
whitenoise==6.9.0
